Move to new pairing.cc file
